update setup token
