+++ src/service/apiService.js
@@ -1,0 +1,5 @@
+import axiosInstance from './axiosService';
+
+export const getUserData = () => {
+  return axiosInstance.get('');
+};
